--- vim/.config/nvim/lua/lsp/lua_ls.lua
@@ -1,9 +1,0 @@
-return {
-    settings = {
-        Lua = {
-            diagnostics = {
-                globals = { "vim", "hs" },
-            },
-        },
-    },
-}
