--- vim/.config/nvim/lua/lsp/ruff.lua
@@ -1,8 +1,0 @@
--- Server-specific settings for ruff.
--- This is loaded by lua/lsp/servers.lua and merged with base configuration.
---
--- Note: The cmd is defined in servers.lua to use project-local .venv/bin/ruff
-
-return {
-    settings = {},
-}
