+++ vim/.config/nvim/lua/lsp/servers/lua_ls.lua
@@ -1,0 +1,23 @@
+local utils = require('lsp.utils')
+
+return vim.tbl_extend("force", utils.base_config, {
+    cmd = { utils.mason_bin .. 'lua-language-server' },
+    filetypes = { 'lua' },
+    root_markers = {
+        '.luarc.json',
+        '.luarc.jsonc',
+        '.luacheckrc',
+        '.stylua.toml',
+        'stylua.toml',
+        'selene.toml',
+        'selene.yml',
+        '.git',
+    },
+    settings = {
+        Lua = {
+            diagnostics = {
+                globals = { "vim", "hs" },
+            },
+        },
+    },
+})
