+++ vim/.config/nvim/lua/lsp/servers/rust_analyzer.lua
@@ -1,0 +1,10 @@
+local utils = require('lsp.utils')
+
+return vim.tbl_extend("force", utils.base_config, {
+    cmd = { utils.mason_bin .. 'rust-analyzer' },
+    filetypes = { 'rust' },
+    root_markers = { 'Cargo.toml', 'rust-project.json' },
+    settings = {
+        ['rust-analyzer'] = {},
+    },
+})
