--- vim/.config/nvim/lua/lsp/yamlls.lua
@@ -1,10 +1,0 @@
-return {
-    settings = {
-        yaml = {
-            -- TODO - dynamically check if schemas are present
-            schemas = {
-                [".vscode/schema.json"] = "**/*.y*ml",
-            },
-        },
-    },
-}
